+++ config/production.json
@@ -1,6 +1,7 @@
 {
   "ORM_CONFIG": {
-    "database": "/data/theta-data/database_v3/",
+    
+    "database": "/data/theta-data/database/",
     "synchronize": true,
     "autoLoadEntities": true,
     "name": "THETA_DATA",
@@ -23,22 +24,7 @@
     "ttl": 60,
     "limit": 60
   },
-  "SMART_CONTRACT_VERIFY_DETECT_TIMES" : 10,
-  "RE_SYNC_BALANCE" : false,
-  "STAKE_ANALYSE_START_HEIGHT" : 14590900,
-  "CONFLICT_TRANSACTIONS" : ["0x4a9c74bf71ac6a3591672d67e975887ef8374ba34b45342305e656666d9cdcf1","0x82236267393c31c17bf4e79fb14f9fc0396c05821ebd357e6b420a43933d64db"],
-  "NFT":{
-    "ANALYSE_NUMBER" : 20,
-    "ANALYSE_INTERVAL" : 1000
-  },
-  "SMART_CONTRACT":{
-    "ANALYSE_NUMBER" : 20,
-    "ANALYSE_INTERVAL" : 1000,
-    "THETA_NODE_HOST" : "https://theta-bridge-rpc.thetatoken.org/rpc",
-    "START_HEIGHT":	14590900
-  },
-  "STAKE":{},
-  "TX":{},
-  "WALLET":{}
+  "SMART_CONTRACT_VERIFY_DETECT_TIMES" : 20,
+  "RE_SYNC_BALANCE" : false
 
 }
