--- nest-cli.smart-contract-analyse.json
@@ -1,5 +1,0 @@
-{
-  "collection": "@nestjs/schematics",
-  "sourceRoot": "src",
-  "entryFile":"analyse/analyse-smart-contract"
-}
