+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "test-theta-data",
-  "version": "0.0.3",
+  "version": "0.0.1",
   "description": "",
   "author": "",
   "private": true,
@@ -15,11 +15,6 @@
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "analyse-smart-contract": "cross-env NODE_ENV=dev nest start -c nest-cli.smart-contract-analyse.json",
-    "analyse-nft": "cross-env NODE_ENV=dev nest start -c nest-cli.nft-analyse.json",
-    "analyse-stake": "cross-env NODE_ENV=dev nest start -c nest-cli.stake-analyse.json",
-    "analyse-tx": "cross-env NODE_ENV=dev nest start -c nest-cli.tx-analyse.json",
-    "analyse-wallets": "cross-env NODE_ENV=dev nest start -c nest-cli.wallets-analyse.json",
     "start": "cross-env NODE_ENV=dev nest start",
     "start:dev": "cross-env NODE_ENV=dev nest start --watch",
     "start:debug": "cross-env NODE_ENV=dev nest start --debug --watch",
@@ -37,6 +32,7 @@
     "@nestjs/throttler": "^2.0.0",
     "@nestjs/typeorm": "^8.0.3",
     "apollo-server-express": "^3.6.3",
+    "await-sleep": "^0.0.1",
     "better-sqlite3": "^7.5.0",
     "cache-manager": "^3.6.0",
     "config": "^3.3.7",
