+++ src/block-chain/smart-contract/smart-contract.module.ts
@@ -6,8 +6,6 @@ import { SmartContractService } from './smart-contract.service'
 import { SmartContractResolver } from './smart-contract.resolver'
 import { NftModule } from './nft/nft.module'
 import { CommonModule } from 'src/common/common.module'
-import { SmartContractAnalyseService } from './smart-contract-analyse.service'
-// import { AnalyseService } from 'src/analyse/analyse.service'
 
 @Module({
   imports: [
@@ -18,7 +16,7 @@ import { SmartContractAnalyseService } from './smart-contract-analyse.service'
     NftModule,
     CommonModule
   ],
-  providers: [SmartContractService, SmartContractResolver, SmartContractAnalyseService],
+  providers: [SmartContractService, SmartContractResolver],
   exports: [SmartContractService]
 })
 export class SmartContractModule {}
