--- src/guard.ts
@@ -1,12 +1,0 @@
-import { ExecutionContext, Injectable } from '@nestjs/common'
-import { GqlExecutionContext } from '@nestjs/graphql'
-import { ThrottlerGuard } from '@nestjs/throttler'
-
-@Injectable()
-export class GqlThrottlerGuard extends ThrottlerGuard {
-  getRequestResponse(context: ExecutionContext) {
-    const gqlCtx = GqlExecutionContext.create(context)
-    const ctx = gqlCtx.getContext()
-    return { req: ctx.req, res: ctx.res }
-  }
-}
